refactor: 내 달력에 추가한 배너 관심사 분리
