+++ src/main/java/com/techeer/fmstudio/domain/banner/controller/WholeBannerController.java
@@ -1,19 +1,13 @@
 package com.techeer.fmstudio.domain.banner.controller;
 
-import com.techeer.fmstudio.domain.banner.domain.MyBannerList;
 import com.techeer.fmstudio.domain.banner.dto.mapper.BannerMapper;
-import com.techeer.fmstudio.domain.banner.dto.request.CustomBannerAddMyBannerRequest;
-import com.techeer.fmstudio.domain.banner.dto.request.MyBannerDeleteRequest;
 import com.techeer.fmstudio.domain.banner.dto.response.BannerPageInfo;
-import com.techeer.fmstudio.domain.banner.dto.response.DeleteResponse;
-import com.techeer.fmstudio.domain.banner.dto.response.MyBannerInfo;
 import com.techeer.fmstudio.domain.banner.service.WholeBannerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
-import javax.validation.Valid;
 
 @RestController
 @RequiredArgsConstructor
@@ -31,33 +25,5 @@ public class WholeBannerController {
         BannerPageInfo foundBannerList = wholeBannerService.getWholeBannerByPagination(page,size);
         return ResponseEntity.status(HttpStatus.OK)
                 .body(foundBannerList);
-    }
-
-    @PostMapping("/mybanners/{id}")
-    public ResponseEntity<MyBannerInfo> addBannerToMyList(
-            @Valid @RequestBody CustomBannerAddMyBannerRequest request,
-            @PathVariable Long id
-    ){
-
-        MyBannerList myBannerList = wholeBannerService.addMyBanner(request, id);
-        MyBannerInfo myBannerInfo = MyBannerInfo.builder()
-                .nickname(myBannerList.getMember().getNickname())
-                .bannerId(myBannerList.getBanner().getId())
-                .build();
-
-        return ResponseEntity.status(HttpStatus.OK)
-                .body(myBannerInfo);
-    }
-
-    @DeleteMapping("mybanners/{bannerId}")
-    public ResponseEntity<DeleteResponse> deleteMyBanner(
-            @Valid @RequestBody MyBannerDeleteRequest request,
-            @PathVariable Long bannerId
-    ){
-
-        wholeBannerService.deleteMyBanner(request, bannerId);
-
-        return ResponseEntity.status(HttpStatus.OK)
-                .body(DeleteResponse.builder().id(bannerId).build());
     }
 }
